Add baseUrl to tsconfig.json (#8)
`baseUrl` is a required field if you use `paths`. I am sure a lot of tools will just assume the `tsconfig.json`'s current directory if it's missing, but it's technically required [as part of the spec](https://www.typescriptlang.org/tsconfig#paths). IntelliJ for one strictly conforms to this, and it won't work correctly without the `baseUrl`.