Fix TypeScript errors in generated project
Previously running `npx tsc` in a freshly generated project would result
in a few type errors.

I added stricter type declarations, especially to `StartContext` from
looking at the existing code.

Additionally, I added a declaration file for the `~/root` file to exist
so that it wouldn't be necessary to use `@ts-expect-error` when
importing it in this library's code. This also fixes an error in the
generated project because `@ts-expect-error` would fail if the file
actually exists and there is no TypeScript error.

One thing I didn't have a good solution for is in the `<Meta>` component
where the `ssr()` function returns `{ t: string }` but the `<Assets>`
component expects a JSX element as its `children`. So, I just added the
`@ts-expect-error` comment there.

There are still some errors when running `tsc` inside the `solid-start`
project though.